Update README.md
Replication of MALL Using C Language

By implementing this simplified version of MALL in C, we aim to provide a foundational understanding of how language learning applications can be built and operated. This project serves as a starting point for further enhancements and experiments in the field of educational technology.